Move back the Constants class.
